refactor: moved all Toolbar options onto TextEditorDrawer
