--- src/components/ToolBar/SheetFormatOptions/Events.ts
@@ -1,7 +1,0 @@
-export const e_pageOrientation = new CustomEvent("page-orientation", {
-  bubbles: true,
-});
-
-export const e_pageMargin = new CustomEvent("page-margin", {
-  bubbles: true,
-});
