+++ src/views/TextEditor/TextEditorDrawer/TextEditorDrawer.vue
@@ -3,6 +3,7 @@ import { computed } from "vue";
 import { ref } from "vue";
 
 import DrawerPageOptions from "./DrawerPageOptions/DrawerPageOptions.vue";
+import DrawerSignPuddleSearch from "./DrawerSignPuddleSearch/DrawerSignPuddleSearch.vue";
 
 const drawer = ref(true);
 const rail = ref(false);
@@ -16,6 +17,7 @@ const chevronDirection = computed(() => {
     :rail="rail"
     permanent
     @click="rail = false"
+    width="320"
   >
     <v-list-item title="LetraSinal">
       <template v-slot:append>
@@ -33,7 +35,7 @@ const chevronDirection = computed(() => {
           <v-list-item
             v-bind="props"
             prepend-icon="mdi-file"
-            title="Página"
+            title="Configurações da Página"
           ></v-list-item>
         </template>
 
@@ -47,9 +49,39 @@ const chevronDirection = computed(() => {
           <v-list-item
             v-bind="props"
             prepend-icon="mdi-hand"
-            title="Sinal"
+            title="Adicionar Sinais"
           ></v-list-item>
         </template>
+
+        <v-list-group value="signpuddle">
+          <template v-slot:activator="{ props }">
+            <v-list-item v-bind="props" title="SignPuddle">
+              <template v-slot:prepend>
+                <v-icon>
+                  <img
+                    class="sp-btn"
+                    width="25"
+                    src="../../../assets/sign-puddle-icon.png"
+                    alt="SignPuddle"
+                  />
+                </v-icon>
+              </template>
+            </v-list-item>
+          </template>
+          <DrawerSignPuddleSearch />
+        </v-list-group>
+
+        <v-list-group value="fsw">
+          <template v-slot:activator="{ props }">
+            <v-list-item v-bind="props" title="Formal SignWriting" disabled>
+              <template v-slot:prepend>
+                <div :style="'margin-right: 1.5rem;'">
+                  FSW
+                </div>
+              </template>
+            </v-list-item>
+          </template>
+        </v-list-group>
       </v-list-group>
 
       <v-divider></v-divider>
